OAK-9802: improved ElasticIndexWriterDisabledTest
